refactor funnel.go
